backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
PyPDF2
pandas
python-docx
